Realised the input/output inverters could be replaced with NAND and multiple cycles, lol
This went from the HACK ALU to almost the MyNOR "ALU". Just an adder, an 8 channel NAND gate, and a multiplexer! The NAND is a free additional output from the adder, too! (It also has AND and XOR out for free, but I am not using those as both are compound gates of NAND, and I save on multiplexers this way.)